requirements.txt
asgiref==3.2.7
awscli==1.18.69
Babel==2.8.0
botocore==1.16.19
colorama==0.4.3
debugger==1.3
dj-database-url==0.5.0
Django==3.0.6
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-nginx-image==0.3.2
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
Pillow==7.1.2
psycopg2==2.8.5
pyasn1==0.4.8
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
rsa==3.4.2
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
